Add new properties to the ConnectionFromClient: 1) The CurrentStatus property allows users to know what each connection is doing at any given time. 2) In case the conn. is idling (waiting for new content of the socket) users can know for how long examining CurrentIdlingTime. 3) In case the conn. is running (executing an actual method) users can know for how long examining CurrentRunningTime. (Maybe here I'm missing TotalIdlingTime and TotalRunningTime) 4) Users can know how many bytes this conn has read and written using TotalBytesRead and TotalBytesWritten. 5) The same goes for the total time spent reading and writing to the network through TotalTimeReading and TotalTimeWritting.
We also have two new members: 'mIdleTimeoutMillis' and 'mRunTimeoutMillis'. When we are going to wait for new content on the socket (await mRpcSocket.WaitForDataAsync) we cancel the token after mIdleTimeoutMillis. When we are going to run the actual code of the RPC invocation we cancel the token after mRunTimeoutMillis.

I'm thinking that at this last point (before running the actual method) we should keep the CancellationTokenSource and monitor somehow the RpcSocket, so in case the socket closes, we cancel the entire method execution.