For each new ConnectionFromClient, we will create a new CancellationTokenSource, then save it in an in-memory collection along with the ConnectionFromClient itself. This way, we can close individual connections at will. There is a background loop that looks for the following every 30 seconds (although it is configurable): 1) Dead connections, to remove them from the collection. 2) Connections that are running a method but whose socket is no longer connected (the connection died while running user code), in order to cancel them and clean them up - this should be configurable though.
Again, we don't need to look for connections that have been idling for too long here because when waiting for new content on the socket we do so using a linked CancellationToken that expires after the ConIdleTimeoutMillis. The same thing goes for connections handling a user method that has been running for too long.