Renamed projects 'dotnetRpc' > 'miloRPC'.
Milo of Croton, a wrestler who has gone down in history for more than his skills on the mat, is also legendary for his superhuman appetite. He won the Olympic wrestling title an incredible six times in a row between 536 and 520 BC, and grabbed the laurel at another 27 championships at the Nemean, Pythian and Isthmian Games as well.